# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Business API",
    "version": "2.0.0",
    "description": "Enterprise-grade middleware API for CAC.gov.ng business registration. All 46 core routes documented."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Use 'Authorization: Token ck_...' for customer API key auth"
      },
      "CustomerApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Customer API key in format: Token ck_your_api_key_here"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/customer/api-keys": {
      "get": {
        "tags": [
          "Customer"
        ],
        "summary": "List API keys",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Keys"
          }
        }
      },
      "post": {
        "tags": [
          "Customer"
        ],
        "summary": "Create API key (returns plaintext once)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/admin/customers/{id}/sync-portal-keys": {
      "post": {
        "description": "Synchronize customer portal keys with database",
        "tags": [
          "Admin - Customer Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Sync portal keys",
        "operationId": "syncCustomerPortalKeys",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Keys synchronized successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      }
    },
...
    "/admin/admins": {
      "post": {
        "description": "Create a new admin user (super admin only)",
        "tags": [
          "Admin - System Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "properties": {
                        "admin": {
                          "properties": {
                            "email": {
                              "type": "string"
                            },
                            "permissions": {
                              "items": {
                                "type": "string"
                              },
                              "type": "array"
                            },
                            "id": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string"
                            }
                          },
                          "type": "object"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Admin created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        },
        "summary": "Create new admin user",
        "operationId": "createAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": {
                    "format": "email",
                    "example": "newadmin@company.com",
                    "type": "string"
                  },
                  "permissions": {
                    "example": [
                      "view_all",
                      "manage_customers"
                    ],
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "password": {
                    "type": "string",
                    "example": "AdminPass123"
                  },
                  "role": {
                    "enum": [
                      "admin",
                      "super_admin"
                    ],
                    "example": "admin",
                    "type": "string"
                  }
                },
                "type": "object",
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          },
          "required": true
        }
      }
    },
    "/admin/customers/{id}/keys/{keyId}": {
      "delete": {
        "description": "Revoke a specific API key for a customer (admin action)",
        "tags": [
          "Admin - Customer Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Revoke customer's API key",
        "operationId": "revokeCustomerKey",
        "parameters": [
          {
            "description": "Customer ID",
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "description": "API Key ID",
            "name": "keyId",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "API key revoked"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Customer or key not found"
          }
        }
      }
    },
    "/admin/customers/{id}/keys": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Admin Generated Key"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "operationId": "generateCustomerKey",
        "responses": {
          "200": {
            "description": "API key generated"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "description": "Create a new API key for a customer (admin action)",
        "summary": "Generate API key for customer",
        "tags": [
          "Admin - Customer Management"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ]
      },
      "get": {
        "description": "Get all API keys for a specific customer",
        "tags": [
          "Admin - Customer Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List customer's API keys",
        "operationId": "listCustomerKeys",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "properties": {
                        "keys": {
                          "items": {
                            "type": "object"
                          },
                          "type": "array"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "List of customer's API keys"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/admin/customers/{id}/suspend": {
      "post": {
        "description": "Suspend a customer account temporarily",
        "tags": [
          "Admin - Customer Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Suspend customer account",
        "operationId": "suspendCustomer",
        "parameters": [
          {
            "description": "Customer ID",
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Customer suspended successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      }
    },
    "/admin/customers/{id}/activate": {
      "post": {
        "description": "Activate a suspended customer account",
        "tags": [
          "Admin - Customer Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Activate customer account",
        "operationId": "activateCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Customer activated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      }
    },
...
  }
}